+++ examples/example5.c
@@ -66,7 +66,7 @@ main (void)
     perror ("fgets");
   buf[strlen (buf) - 1] = '\0';
 
-  printf ("Read string (length %ld): ", strlen (buf));
+  printf ("Read string (length %ld): ", (long int) strlen (buf));
   for (i = 0; i < strlen (buf); i++)
     printf ("%02x ", buf[i] & 0xFF);
   printf ("\n");
@@ -87,7 +87,7 @@ main (void)
       return 2;
     }
 
-  printf ("ToASCII string (length %ld): %s\n", strlen (p), p);
+  printf ("ToASCII string (length %ld): %s\n", (long int) strlen (p), p);
 
   rc = idna_to_unicode_8z4z (p, &r, 0);
   free (p);
@@ -107,7 +107,7 @@ main (void)
   free (r);
   if (rc == TLD_INVALID)
     {
-      printf ("Domain rejected by TLD check, Unicode position %ld\n", errpos);
+      printf ("Domain rejected by TLD check, Unicode position %ld\n", (long int) errpos);
       return 1;
     }
   else if (rc != TLD_SUCCESS)
